feat(buf-linter): fixing linter errors
