+++ Makefile
@@ -102,7 +102,7 @@ protoc: ## Run protoc
 	@echo "==> Running protoc"
 	protoc --go_out=. --go_opt=paths=source_relative \
         --go-grpc_out=. --go-grpc_opt=paths=source_relative \
-        protos/helloworld/helloworld.proto
+        protos/helloworld/v1/helloworld.proto
 
 help: echo-env
 	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-30s\033[0m %s\n", $$1, $$2}'
