--- protos/helloworld/helloworld.proto
@@ -1,24 +1,0 @@
-syntax = "proto3";
-
-option go_package = "protos/helloworld";
-
-service HelloWorld {
-  rpc SayHelloSimple(Message) returns (ReplyMessage) {}
-  rpc SayHelloServerStreaming(Message) returns (stream ReplyMessage) {}
-  rpc SayHelloClientStreaming(stream Message) returns (ReplyMessages) {}
-  rpc SayHelloBothStreaming(stream Message) returns (stream ReplyMessage) {}
-}
-
-message Message {
-  string text = 1;
-}
-
-message ReplyMessage {
-  string text = 1;
-}
-
-message ReplyMessages {
-  int64 startTime = 1;
-  int64 endTime = 2;
-  repeated Message messages = 3;
-}
